BackEnd/requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Cors==3.0.10
flask-request-validator==4.1.0
httplib2==0.19.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauth2client==4.1.3
openpyxl==3.0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
rsa==4.7.2
six==1.16.0
Werkzeug==1.0.1
XlsxWriter==1.4.3
boto3~=1.17.82